refactor(audio): move malgo code to subpackage; delete stub_nocgo
internal/audio kept a parallel stub_nocgo.go that hand-mirrored the
cgo half — the BackendFactory interface, AudioData field layout, and
DecoderRegistry method set diverged silently across the two universes.
The CLI called GetSupportedBackends, which existed only cgo-side, so
CGO_ENABLED=0 go build ./internal/audio/... failed.

Move malgo-dependent files (context.go, decoder.go, registry.go, the
three decoder.go files, playback.go, malgo_backend.go -> backend.go,
and their _test.go siblings) into internal/audio/malgo with //go:build
cgo. internal/audio becomes cgo-free: backend.go (AudioBackend
interface), source.go (AudioSource + FilePather + FileSource +
ReaderSource), platform.go, system_command_backend.go, factory.go.

Wiring uses a registration pattern (audio.RegisterBackend) so the
audio package never imports the malgo subpackage — the malgo
subpackage's init() registers a constructor under "malgo". The CLI
blank-imports the subpackage via cli_cgo.go //go:build cgo so the
init() fires only when cgo is on. Under !cgo the subpackage doesn't
compile and the registration is absent; NewBackend("malgo") returns
ErrBackendNotAvailable with a clear "build with cgo?" message.

Source-side decoupling: FileSource no longer holds *DecoderRegistry —
decoding lives entirely inside the malgo subpackage. FileSource now
just holds a path and exposes a naive FormatHint() based on
filepath.Ext for SystemCommandBackend's temp-file fallback.

Stub_nocgo.go is deleted; factory_nocgo.go is deleted (its job is
done by the now-cgo-free factory.go). Phase D regression tests
(atomic volume, sync.Once context, deviceEntry sync.Once Uninit,
IsPlaying derivation) all moved with playback.go to the subpackage;
-race -count=3 against the new location PASSES. CGO_ENABLED=0
go build ./... now succeeds.

Resolves review finding #5.
