+++ internal/audio/malgo/context_test.go
@@ -1,4 +1,4 @@
-package audio
+package malgo
 
 import (
 	"testing"
