+++ internal/audio/malgo/decoder_ctx_test.go
@@ -1,6 +1,6 @@
 //go:build cgo
 
-package audio
+package malgo
 
 import (
 	"bytes"
