+++ internal/audio/malgo/mp3_decoder.go
@@ -1,6 +1,6 @@
 //go:build cgo
 
-package audio
+package malgo
 
 import (
 	"context"
